requirements.txt
flask
flask-cors
sentence-transformers
supabase
python-dotenv
requests
torch
transformers
huggingface-hub
